More improvements on the subscriber builder API
Prior to this commit, the `Subscriber` itself was still responsible
for accepting some configuration parameters, such as the cancellation
token or the database pool.

This commit finally moves all of such responsabilities to the
`Builder` module itself. For that, we use the "type-safe builder API
pattern", which guarantees that all required parameters (for now
only the database pool) are provided.

The type-safe API also ensures a well-defined order of configurations
in the builder.