Introduce new error types
This commit introduces the `InternalError`, `JobExecutionError`,
`JobPublishError` and `SetupError` error types, which are used to
represent different error scenarios that may happen during Fila's
operation.

Such a distinction (esp. the `JobExecutionError`) is important as
we'll soon introduce a callback-based error-handling mechanism to
deal with job execution errors, and having separate error types
simplifies the handler's code since it won't have to bother with
maching on unrelated error variants (for example, the `SetupError`
cases).